+++ src/utils/utils.ts
@@ -1,19 +1,36 @@
 import * as vscode from 'vscode';
 import * as childProcess from 'child_process';
 import { promisify } from 'util';
-import { ObjectEncodingOptions } from 'fs';
-import { privateEncrypt } from 'crypto';
+import * as fs from 'fs';
 
 const execAsync = promisify(childProcess.exec);
 
-export async function executeCmd(cmd: string): Promise<[string, string]> {
-    const options: ObjectEncodingOptions & childProcess.ExecOptions = {
+export async function executeCmd(cmd: string, cachePath?: string, cacheTimeout?: number): Promise<[string, string]> {
+    if (cachePath && fs.existsSync(cachePath)) {
+        const jsonData = fs.readFileSync(cachePath, 'utf8');
+        const saveObject = JSON.parse(jsonData);
+        const now = Date.now();
+        if (now - saveObject.time < (cacheTimeout ?? 1000) && cmd === saveObject.cmd) {
+            return [saveObject.out.trim(), saveObject.err ? saveObject.err.trim() : ''];
+        }
+    }
+    const options: fs.ObjectEncodingOptions & childProcess.ExecOptions = {
         encoding: 'utf8',
         timeout: 30000,
         maxBuffer: 1024 * 1024,
     };
     try {
         const { stdout, stderr } = await execAsync(cmd, options);
+        if (cachePath) {
+            const saveObject = {
+                'time': Date.now(),
+                'cmd': cmd,
+                'out': stdout,
+                'err': stderr,
+            };
+            const jsonData = JSON.stringify(saveObject);
+            fs.writeFileSync(cachePath, jsonData, 'utf8');
+        }
         return [stdout.trim(), stderr ? stderr.trim() : ''];
     } catch (error) {
         return ['', `${error}`];